requirements.txt
defusedxml==0.7.1
fonttools==4.57.0
fpdf2==2.8.3
numpy==2.2.5
pandas==2.2.3
pillow==11.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
